refactor using a wrapper
