+++ components/api/pageHeader/index.js
@@ -1,0 +1,4 @@
+import PageHeaderWrapper from './pageHeaderWrapper';
+import PageHeader from './pageHeader';
+
+export { PageHeaderWrapper, PageHeader };
